fix: Fix lib compilatioon
