+++ espflash/src/lib.rs
@@ -56,8 +56,6 @@ pub mod command;
 pub mod connection;
 pub mod elf;
 pub mod error;
-#[cfg(feature = "serialport")]
-#[cfg_attr(docsrs, doc(cfg(feature = "serialport")))]
 pub mod flasher;
 pub mod image_format;
 #[cfg(feature = "serialport")]
