+++ espflash/src/targets/flash_target/mod.rs
@@ -2,12 +2,15 @@ use bytemuck::{Pod, Zeroable};
 
 pub(crate) use self::ram::MAX_RAM_BLOCK_SIZE;
 pub use self::{esp32::Esp32Target, esp8266::Esp8266Target, ram::RamTarget};
-use crate::{connection::Connection, elf::RomSegment, error::Error, flasher::ProgressCallbacks};
+#[cfg(feature = "serialport")]
+use crate::connection::Connection;
+use crate::{elf::RomSegment, error::Error, flasher::ProgressCallbacks};
 
 mod esp32;
 mod esp8266;
 mod ram;
 
+#[cfg(feature = "serialport")]
 /// Operations for interacting with a flash target
 pub trait FlashTarget {
     /// Begin the flashing operation
